Simplify PermutationTable
For some reason, PermutationTable was jumping through some really over-
complicated hoops, implementing the Distribution<PermutationTable> trait
on Standard so it could generate the table through rng.gen(), when all
it actually needed to do was build an array of 256 shuffled numbers. So
now it just does that, the simple way.

Also, it now uses SmallRng, which unlike the XorShiftRng that it used
before, will still produce good randomness with very small seed values,
like 0 or 1. Previously, a lot of the noise functions had clear
repeating patterns when using those small seed values. They look better
now.

+++ src/noise_fns.rs
@@ -50,8 +50,8 @@ where
 /// Trait for functions that require a seed before generating their values
 pub trait Seedable {
     /// Set the seed for the function implementing the `Seedable` trait
-    fn set_seed(self, seed: u32) -> Self;
+    fn set_seed(self, seed: u64) -> Self;
 
     /// Getter to retrieve the seed from the function
-    fn seed(&self) -> u32;
+    fn seed(&self) -> u64;
 }
