+++ src/noise_fns/generators/fractals.rs
@@ -20,14 +20,14 @@ pub trait MultiFractal {
     fn set_persistence(self, persistence: f64) -> Self;
 }
 
-fn build_sources<Source>(seed: u32, octaves: usize) -> Vec<Source>
+fn build_sources<Source>(seed: u64, octaves: usize) -> Vec<Source>
 where
     Source: Default + Seedable,
 {
     let mut sources = Vec::with_capacity(octaves);
     for x in 0..octaves {
         let source = Source::default();
-        sources.push(source.set_seed(seed + x as u32));
+        sources.push(source.set_seed(seed + x as u64));
     }
     sources
 }
