+++ src/noise_fns/generators/fractals/hybridmulti.rs
@@ -38,19 +38,19 @@ pub struct HybridMulti {
     /// persistence produces "rougher" noise.
     pub persistence: f64,
 
-    seed: u32,
+    seed: u64,
     sources: Vec<Perlin>,
 }
 
 impl HybridMulti {
-    pub const DEFAULT_SEED: u32 = 0;
+    pub const DEFAULT_SEED: u64 = 0;
     pub const DEFAULT_OCTAVES: usize = 6;
     pub const DEFAULT_FREQUENCY: f64 = 2.0;
     pub const DEFAULT_LACUNARITY: f64 = core::f64::consts::PI * 2.0 / 3.0;
     pub const DEFAULT_PERSISTENCE: f64 = 0.25;
     pub const MAX_OCTAVES: usize = 32;
 
-    pub fn new(seed: u32) -> Self {
+    pub fn new(seed: u64) -> Self {
         Self {
             seed,
             octaves: Self::DEFAULT_OCTAVES,
@@ -99,7 +99,7 @@ impl MultiFractal for HybridMulti {
 }
 
 impl Seedable for HybridMulti {
-    fn set_seed(self, seed: u32) -> Self {
+    fn set_seed(self, seed: u64) -> Self {
         if self.seed == seed {
             return self;
         }
@@ -111,7 +111,7 @@ impl Seedable for HybridMulti {
         }
     }
 
-    fn seed(&self) -> u32 {
+    fn seed(&self) -> u64 {
         self.seed
     }
 }
