+++ src/noise_fns/generators/open_simplex.rs
@@ -11,14 +11,14 @@ use crate::{
 /// Noise function that outputs 2/3/4-dimensional Open Simplex noise.
 #[derive(Clone, Copy, Debug)]
 pub struct OpenSimplex {
-    seed: u32,
+    seed: u64,
     perm_table: PermutationTable,
 }
 
 impl OpenSimplex {
-    const DEFAULT_SEED: u32 = 0;
+    const DEFAULT_SEED: u64 = 0;
 
-    pub fn new(seed: u32) -> Self {
+    pub fn new(seed: u64) -> Self {
         Self {
             seed,
             perm_table: PermutationTable::new(seed),
@@ -34,7 +34,7 @@ impl Default for OpenSimplex {
 
 impl Seedable for OpenSimplex {
     /// Sets the seed value for Open Simplex noise
-    fn set_seed(self, seed: u32) -> Self {
+    fn set_seed(self, seed: u64) -> Self {
         // If the new seed is the same as the current seed, just return self.
         if self.seed == seed {
             return self;
@@ -47,7 +47,7 @@ impl Seedable for OpenSimplex {
         }
     }
 
-    fn seed(&self) -> u32 {
+    fn seed(&self) -> u64 {
         self.seed
     }
 }
